test(fakes): merge aptos_chain_methods_test; drop redundant lifecycle/shim tests
Consolidate all FakeAptosChain tests into aptos_chain_test.go. Remove
TestFakeAptosChain_Lifecycle (subsumed by InitialiseStartsService) and
TestFakeAptosChain_InterfaceShims (covers no-op stubs, tests nothing real).
